--- examples/apps/presence-tracker/src/index.ts
@@ -1,8 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export { FocusTracker, IFocusSignalPayload, IFocusTrackerEvents } from "./FocusTracker";
-export { ITrackerAppModel, TrackerContainerRuntimeFactory } from "./containerCode";
-export { renderFocusPresence, renderMousePresence } from "./app";
