+++ examples/apps/presence-tracker/webpack.test.js
@@ -33,7 +33,7 @@ module.exports = (env) => {
 			library: "[name]",
 			// https://github.com/webpack/webpack/issues/5767
 			// https://github.com/webpack/webpack/issues/7939
-			devtoolNamespace: "fluid-example/draft-js",
+			devtoolNamespace: "fluid-example/presence-tracker",
 			libraryTarget: "umd",
 		},
 		devServer: {
